LPS-66452 Create base class for regexp based matcher
It is expected that most implementations will be based on regular
expressions, so it is useful to abstract all the low level machinery
into an abstract base class.
